Exception configured

+++ src/main/java/br/janioofi/vuttr/domain/services/ToolsService.java
@@ -3,18 +3,16 @@ package br.janioofi.vuttr.domain.services;
 import br.janioofi.vuttr.domain.DTO.ToolsDto;
 import br.janioofi.vuttr.domain.entities.Tools;
 import br.janioofi.vuttr.domain.repositories.ToolsRepository;
+import br.janioofi.vuttr.exceptions.RecordNotFoundException;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
+@RequiredArgsConstructor
 public class ToolsService {
-
     private final ToolsRepository repository;
-
-    public ToolsService(ToolsRepository repository) {
-        this.repository = repository;
-    }
 
     public List<Tools> findAll(){
         return repository.findAll();
@@ -34,6 +32,15 @@ public class ToolsService {
     }
 
     public void deleteById(Long id){
+        if(repository.findById(id).isEmpty()){
+            throw new RecordNotFoundException(id);
+        }
         repository.deleteById(id);
+    }
+
+    public Tools findById(Long id){
+        return repository.findById(id).orElseThrow(() -> {
+            throw new RecordNotFoundException(id);
+        });
     }
 }
